WIP Split Node table in three
 - Node: Common config fields. No fields but the autoincremental PK for
   now.

 - NodeExternal: Config for external nodes. Has user set fields

 - NodeDetails: Cache for state of nodes

The latter two have `id` fields, which point to `Node`, and data fields,
which store all the other data. This mimicks the ergonomics of
auto-incremented fields, and allows keeping the data inside the app,
which typically involves the config and details (and config including
`NodeInternal` when it is added) from including redundancies.

This supports adding a `NodeInternal` table, which plays the same role as
`NodeExternal` but for Kiln-managed nodes.