Simplify the migrations relating to the spit node table
The old logic was falled in that if nothing exists, the table creations
would fail due to a foreign key constraint referencing uncreated
tables.

The solution is simply to move the table creations behind the same
guard for the migration that actually moves the data. If nothing exists,
the automigrate does everything. If the tables are all in the right
state, nothing happens. If the old node table exists, then the table
creations are done (with the foreign key constraints valid referring to
existing tables) and and the data moved.