tools for RPC+database interaction
They're good for the case where you're doing a database
transaction that is relatively small and needs a relatively
small amount of *cacheable* data from the node, but there
are a lot of internal dependencies between queries.  With a
little work they could also be good for the case where there's
a lot of data to query from the node and a lot of dependencies,
but the dependencies are not deep.  With some more work, could
become good for almost all cases, with the aid of some manual
tuning to indicate good checkpoints and data bottlenecks.