Fix tooltip, again
Reverts the overflow change which allowed the tooltip to appear outside
of the container: it breaks the fixed sidebar. Also turns out when you
use overflow-x visible and overflow-y auto, the visible bit is
quietly set to auto, so you can't fix it that way. Instead I just
shifted the tooltip across so it doesn't overlap.