Fix the last commit
Part of me would rather coalesce,  but it's probably unnecessary.

It turns out that I'd have to write

SET fitness = COALESCE("BlockShellIndex".fitness, EXCLUDED.fitness),  ...

which is a bit longer, which I wouldn't care about if we were using
prepared statements.  It'd be nice to have a better story for prepared
statements.