feat: Babylon and Athens dual support
This adds in functionality to have our RPCs parse either V004 or V005
versions of things for the JSON so that we can properly index blocks
that are across both protocol versions.

It does this by versioning the types in TBM-lib into V004 and V005. If
you use the NodeRPC of V005 you'll be able to get both V004 and V005
blocks out. Where there is a choice, you get a union type at the top
level. Currently the parsing is done with the Parser alternative
instance, which is kinda dodgy but works for the differences that
we have now.

There is no "cross compat" api in the binary formats. Users will need
to explicitly choose a type.

This also adds some tests, which felt important to get this right.

See https://tezos.gitlab.io/master/protocols/005_babylon.html#changes-to-rpcs
for the official changelog between Athens and Babylon.