fix: Makes the protocol switch test work for Babylon
This removes the duplicate proposals because that was blocking the
testing_phase voting. Probably better in the long run to just do less
of them than 0 of them.

Also changes the code submitted to be BABY5H not BabyM1. This means the
tests don't succeed because it's waiting for BABY5H but only sees BabyM1.