fix: Adds a resetLedgerQueue to be run on kiln startup
Before this patch, in certain circumstances, kiln could be killed just
as one of the LedgerAccount "shouldXXX" actions be set, meaning that when
kiln was restarted the action could happen against the ledger (which should
not happen as the UI interaction is gone at that point). This flushes all
of these field to the empty "do nothing" state upon startup.