feat: Adds babylon support, drops alphanet and improves Cache_NoSuitableNode errors dramatically
This adds babylonnet support and drops alphanet support to catch up to
the latest state of affairs.

In doing this I hit a pretty serous issues with my setup (without the OS
public node) where my node had a savepoint that was later than the blocks
that needed to be indexed on a cold cache, so everything got terribly stuck.
Hence, I added some additional error logging for this case so it is obvious
why we aren't hitting the nodes that we are skipping over.