fix: Makes sure the baker rights workers gets all rights
Previous to this patch the baker rights worker was only going
PRESERVED_CYCLES*BLOCKS_PER_CYCLES ahead of the start of this cycle.

It misses an entire cycle of available rights, and worse still the
ViewSelector for baker rights looks for max baker rights progress
of to the end of CURRENT_CYCLE + PRESERVED_CYCLES to show WaitingForRights
instead of GatheringData, so a baker that has no rights just sits Gathering
data till the rights pop up.