feat: Adds the tooltip for curve params on ledger select
This adds a tooltip to each pkh so that you can figure out where it came
from.

The debounce was pretty bad behaviour because it meant that the popover
took 750 ms to disappear which was really awful if you hovered you mouse
upward over all of them in a fast manner.

This throttle change actually seems to work nicer every tooltip so I just
changed it everywhere. In this sense, we don't actually need the tooltip
config changes, but we may as well leave it here just in case we need to
tweak things later in case the throttle vs debounce doesn't work out somewhere.