fix: Nukes out any trace of TzScan from the DB
On an existing setup that had TzScan before, if you don't nuke the TzScan
from the db the view selector crashes with a Prelude.read: no parse error.

There are so many uses of read in Backend.Schema that I'm surprised that
we never hit this before. But I guess we probably don't remove constructors
all that often!