Clarify VirtualBox install instructions.
The existing instructions were vague; the presence of a link to the VirtualBox
website may appear to indicate to a user that they should install the
application from that site.  We have encountered problems with that particular
installation method and therefore the instructions have been updated to clarify
that using the version packaged in the normal OS package repository is better.