Update procedure changed
There are three important parts here

In backend/src/Backend/Upgrade.hs, there is function called
upstreamGitLab which constructs the correct url to fetch the
backend.cabal file from the latest commit on a specified branch. The
backend version (which is also kiln's version) is parsed from this
file. Previously, the url was pointing at obsidian's repo. Now it
points to ours (changed the project id number).

The changelog url has also been changed to point to our url.

The last change is that the requestHandler in Backend.hs has been
changed so that the upgradeBranch (the branch needed to get the
version mentioned above) now points to "develop" instead of "master".
If this needs to be changed to something else, we should iron out that
issue.