Revise throwing function in Node.hs
This function would throw an error on bad node rpc calls. Generally
this is fine until a rpc call fails because of bad json that is also
unfortunately rather long. If it is too long, the json output can
swarm the console and prevent someone from diagnosing the underlying
problem. This been revised so that the underlying problem with the
json is still printed without the entirety of the json. When someone
wants to see the json, they can look for it in the logs under the
namespace "kiln-debugging". They just need to set up the filter.