Add PublicRequest_ShowLedgerBatch constructor
As the name suggests, the frontend can use this constructor to ask for
balances associated to a batch of accounts in one-shot instead of
waiting for the TezosClient worker to pick up each account one by one
and then process them.

As a first pass, we serialize account balance requests. Upon further
review, we will consider parallel requests.