write validateNodeConfigFile function
This function validates that the required fields are in the provided
JSON object (for running a custom network). If the user wishes to run a
node on either mainnet, or other established networks, this function
instead receives a Haskell data type corresponding to a well formed
configuration. As such, this function immediately validates that configuration.