[#50] Disable manual fee during baker registration
Problem: If baker key wasn't revealed to the blockchain yet, the
registration step will fail because it's impossible to use custom fee
with the unrevealed baker key.

Solution: Remove an ability to provide manually selected fee for the
baker registration operations because 'tezos-client' itself is capable
of calculating such fee, moreover, the current minimal suggested
fee is much larger that the fee automatically calculated by the
'tezos-client'. To conclude, it doesn't make much sense to let user
choose this operation fee, since it doens't really affect anything and
complicates the process of baker setup.