[#50] Add migration for "LedgerAccount" table
Problem: This MR adds changes to the "LedgerAccount" table scheme and
thus we need a migration, so that it's possible to upgrade the database
from the older versions.

Solution: Add a migration that populates "shouldRegister" column basing
on the existing data from the "shouldRegisterFee" column and drops
"shouldRegisterFee" after that.