[#4] Use timeout in 'chunkedNodeQuery'
Problem: 'chunkedNodeQuery' can block the thread which calls it in case
there is some connection issues to the node, as a result node processing
will stuck and the kiln will show outdated information.

Solution: Limit 'chunkedNodeQuery' execution via timeout. The current
timeout is 2 minutes. Such timeout won't cause constant restarts due to
some blockchain delays on the mainnet (where block interval is 1
minutes) and it's also not that large to be annoying.