[#5] Gather baking and endorsing rights in batches
Problem: Gatherin baker data takes large amount of time. On mainnet
baker worker has to do 5 * 4096 * 2 RPC calls to gather all baking and
endorsing rights for the currently preserved cycles which takes A LOT of
time.

Solution: Gather baker data in levels batches, so that we gather baking
or endorsing rights for 50 levels in a single RPC call.