[#57] Don't create releases for release candidates
Problem: Gitlab doesn't have a feature for creating pre-releases, so
currently, a new release is created for each release candidate which can
be confusing when we point someone to the latest release.

Solution: Don't create a release and don't upload release artifacts to the
gitlab registry when the tag has 'rc' in its name.