[#73] Cache less old rights
Problem: Currently Kiln tries to cache all baking and endorsing rights
for all blocks in 'preserved_cycles' in the past. This takes quite a lot
of time and consumes significant amount of disk space. However, users
usually aren't interested in such old information.

Solution: Add a new option which allows to define the amount of blocks
in the past from which baking and endorsing rights are gathered. By
default this value is 720 blocks (which is equivalent to ~6 hours on
mainnet).