[#69] Get rid of BlockTodo table
Problem: BlockTodo table grows over time and never gets cleaned up, as a
result, it consumes a significant amount of disk space even though it
stores only data about the old blocks that we're not interested in.

Solution: The main BlockTodo table usage is parsing blocks in order to
find accusations.

Updated 'blockWorker' does the following:
* Stores only fixed amount of blocks in the past (720 blocks
  specifically)
* Fetches predecessors hashes, gathers block info, and parses accusations
  in them in one take. In order to avoid excessive work, we store hashes
  for blocks in which accusations were successfully parsed, here we again
  store fixed amount of hashes in the past.