[#76] Add an ability to download snapshot by Kiln
Problem: Currently, it's required to download snapshot and upload it to
the Kiln via UI, this produces an unecessary copy of the snapshot data.
Additionally, there were some issues with the large files upload in the
browser.

Solution: Add additional option for the node setup process. Now it's
possible to provide an URL to the snapshot file. Kiln will download and
import provided snapshot without storing the snapshot file twice.