Enlarge custom chain_id priority
Problem: Currently expected chain_id is taken from the public node, if
the chain_id isn't specified explicitely, the mainnet is taken, this
leads to a 'Node on a wrong network' errors when running Kiln with
custom config network.

Solution: When custom config is provided, the chain_id calculated from
the custom config has the larger priority that the config from the
public node.