Always run binaries workers
Problem: Some of the binaries workers are run only when the name of the
current chain is one of known or there are manually defined paths to the
binaries. However, we want to test some scenarios using custom networks
and existing built-in Tezos binaries.

Solution: Always all binaries workers providing to them optional custom
paths to the binaries.