[#90] Check snapshot file permissions
Problem: When user runs Kiln on macOS and tries to import
snapshot from file path, he wil get 'Invalid snapshot' error
if the snapshot is stored in Desktop, Downloads or Documents
folder, because macOS restricts application access to them.
This error is misleading, and we need to show the correct
error message in this case.

Solution: Add check for read permissions to the snapshot
file and throw corresponding error in case of permissions issue.