[#85] Don't use in-memory RPC responses cache
Problem: Currently Kiln keeps in-memory cache with all recent RPC
responses. As a result, this may lead to an unexpectedly high memory
usage when Kiln performs a lot of RPC requests within short period of
time (e.g. when bootstrapping a node).

Additionally, this cache doesn't make a lot of sense since blocks are
changing quite frequently and cache hits are relatively rare.

Solution: Get rid of this in-memory cache. The only place that now
preserves results of RPC responses is 'NodeQueryT' where it's required
to have these responses while interacting with the DB.
Also, get rid of in-DB responses cache since it's no longer used by the
code.