[#85] Get rid of 'CachedHistory'
Problem: We'd like to reduce the amount of memory that Kiln uses in
runtime. One of things that resides in RAM is 'CachedHistory' which is
constant in size but still consumes significant amount of memory (around
100M, I believe). However, majority of its functionality can be
reimplemented using RPC requests.

Solution: Get rid of 'CachedHistory' and reimplement its usage via node
RPC requests.