[#93] Always get voting period kind from current head
Problem: For some reason, the implementation takes the period kind of
the current head block in case this head isn't the last block in the
period, otherwise, it takes predecessor of the current period kind.
As a result, it leads to a weird voting notifications, where some of the
periods are mentioned twice.

Solution: I don't see any reasons why the period of the head predecessor
should be used, so rework the code to always use the period kind of the
current head block.