[#94] Don't store 'insufficientFunds' flag
Problem: When the proposal voting alert is fired, two transactions to the DB
may conflict due to changes in the 'BakerDaemonInternal' table, which
isn't really affect Kiln functionality, but still produces a lot of
noise in the logs with DB transaction conflicts.

Solution: Use another way to determine whether the baker has sufficient
amount of funds in 'getBakerAddresses': count amount of unresolved
'ErrorLogInsufficientFunds' for the given baker and report that it has
insufficient funds to stake in case there are some unstopped entries in
the 'ErrorLog'.