[#97] Create data directory when create internal node with p2p mode
Problem: Along with the removal of Kiln node, its
data directory is also removed. But when creating a Kiln
node, it's created only in cases of snapshot import. So if
user recreates Kiln node in p2p mode, then an error will occur,
and the node will stuck in 'Initializing' state.

Solution: Create data directory also in case of p2p mode.