[Chore] Reduce timeout in 'obtainLock'
Problem: When user restarts Kiln, he can see the
'internalnode LOCK HELD' in logs, which means that
some process still holds some lock, and new one can start.
The current wait time is 5 mins, which is quite long.

Solution: Ideally, we need to reconsider usage of 'backend'
column in 'ProcessData' and make sure we need this lock.
As a temporary solution, the timeout is reduced to 30 seconds.