Enhance ledger disconnection alerts
Problem: Currently ledger disconnection alert will disappear when user
will open a Tezos Wallet app. However, the baker and endorser daemons
won't be able to bake and endorse, while the UI won't anything about it.

Solution: Display an alert when ledger is connected and Wallet app is
opened on it. This alert informs that one should open Baking app in
order to bake and endorse successfully.