Clear disconnection alerts after successful bake/endorsement
Problem: In order to avoid blocking ledger connectivity checks aren't
performed when baker is two levels within his next opportunity. This may
lead to a situation when there are no connectivity checks for a long
time. As a result, we may incorrectly show that ledger is disconnected
even though Kiln successfully endorses/bakes

Solution: Successful bake/endorsment by an internal baker is an evidence
that ledger is connected, thus we may remove ledger disconnection alert
if we encounter successful bake/endorsement on the block that was after
the ledger disconnection.