[#100] Remove 'tezos-endorser' daemon on Ithaca
Problem: there is no `tezos-endorser` binary for Ithaca
protocol. Now Kiln assumes that there is always both baker
and endorser binaries for each protocol and run both daemons.
We need not to run some of binaries it's not specified in the
paths.

Solution: Added check for the binary path existence, and the
process doesn't start if there is no path to the corresponding
binary.