Consider 'BakerRightsProgress' before ledger connectivity check
Problem: Currently ledger connectivity check isn't performed when the
next baking/endorsing opportunity is two levels around the current head
block. However, in case we haven't gathered rights up to the current head
level, there is still a possibility to run a connectivity check when
baker is supposed to bake/endorse, which may cause missing
bake/endorsement.

Solution: Check that we gathered enough rights before doing the ledger
connectivity check. This way we won't trigger connectivity check when
there are unknown opportunities that are close to the current head.