[#107] Implement alert for missed endorsement bonus
Problem: In Ithaca, there are two delegate roles that may get
some or all of the baking rewards: block proposer and block baker.
Typically proposer and baker are the same, but not necessarily.
We need report to user if proposer and baker are not the same,
i.e when proposer missed the bonus.

Solution: Implemented new alert 'ErrorLogBakerMissedEndorsementBonus'
which is reported for a baker if both conditions below are true:
  * block.proposer == baker
  * block.proposer /= block.baker