[#111] Don't store aeson-encoded values in 'CacheBakingRights' table
Problem: We used to store aeson-encoded sequence of baking
rights for each level in 'CacheBakingRights' table. Now baking
rights cache is priority-aware, and since '(level, priority)'
defines a single unique value, it makes no sense to store
encoded list in 'CacheBakingRights' table, because it always
will be a singleton list.

Solution: Add 'delegate' and 'estimatedTime' columns to
'CacheBakingRights' table and removed 'result' column which
was an encoded list of rights.

Moreover, in order not to complicate the caching logic, it was
decided cache only baking rights with 'priority = 0'. We need
baking rights with different priorities only when we work with
accusations, so since this is a rather rare event we don't use
cache for such rights.