[Chore] Check ledger after disconnection when there is a baker
Problem: currently the ledger connectivity check can result in an
incorrect report of disconnection in cases where the ledger device
is busy, because 'tezos-client' replies with an empty list.
This will result in a state in which Kiln thinks that the ledger is
disconnected and will not check the connection again (as long as there
are endorsing opportunities nearby).

Solution: take the previous connection status into account when
deciding to do (or not) a new connection check.
When there is a baker, the ledger was previously disconnected and
some time has passed, the check will be made again.