[Chore] Set 'max_round' param to be always 0 in 'NodeRPC'
Problem: Since [#111] we use 'max_round' query parameter
in '/helpers/baking_rights' query. We set it to 0 almost
everywhere except accusation handling.

Now we don't use baking rights to detect accusations,
so it's not needed to have an ability to fetch baking
rights with different 'max_round' param, because our
caching logic relies on the invariant that it stores
rights with 'round = 0' and having an ability to fetch
baking rights with different round might be error prone
in the future.

Solution: Remove 'max_round' param from 'NodeQuery*_BakingRights'
data types and set it to 0 when calling RPC.