[#129] Fix 'level -> cycle' calculations in amendment tile
Problem: Kiln shows incorrect cycles info on the amendment
tile and voting popup on mainnet because `level -> cycle`
calculations are incorrect. They use `blocks_per_cycle`
constant from the current protocol, but this constant
was changed since `granada` and may also be changed
in the further protocol.

Solution: Use `latestHead` for which we know the level,
cycle and cycle position to calculate start/end cycles
of amendment period based on this information.