[Chore] Fix baker healthcheck
Problem: There is a bug in baker healthcheck work:
the baker should be marked as unhealthy if it misses
5 enodrsements (or single bake) in a row and then
become healthy after first successful endorsement.
But now it isn't marked as unhealthy in case of 5
or more endorsement misses.

The cause is that 'bakerWorker' does the 'notify' call
on 'BakerDetails' in 2 places:

1. While collecting info about missed rights
2. While updating the information about delegates

The former updates 'missedRightsInRow' counter and makes
notify with the correct 'BakerDetails' data, but the latter
doesn't know anything about 'missedRightsInRow' counter value
and uses default (0) value. So the frontend gets the correct
information about missed rights count first, and then always
get zero value, which leads to the behavior when baker is always
healthy on UI.

Solution: Fetch current 'missedRightsInRow' value from db
along with baker ids in 'updateDelegateDetails' and use it in 'notify'.