[#134] Pass default liquidity baking argument to 013 baker
Problem: There is a new required argument in 'tezos-baker-013-PtJakart':
'--liquidity-baking-toggle-vote'. In case when this argument isn't
provided to baker binary, the binary will fail. We need to provide
some default value for this argument to have an ability to run
baker on 'jakartanet'. The full support of this option will be
added in #135.

Solution: Always pass '--liquidity-baking-toggle-vote pass' argument
to 'tezos-baker-013-PtJakart'.