[#136] Make baker to restart with different liquidity baking option
Problem: '--liquidity-baking-toggle-vote' baker argument is required
since Jakarta. If it hasn't been previously configured by the user,
we set 'pass' value as a default one. However, we need to provide
an opportunity to change this option on already created Kiln baker.

Solution: Add an ability to change '--liquidity-baking-toggle-vote'
option for running Kiln baker, which will restart the baker with
new option value.