[#143] Refactor 'tezosClientWorker' queue
Problem: Currently, tezos-client calls are queued via DB. This approach
leads to some issues with ledger interactions.

Solution: Since the goal of interaction with the ledger is to avoid
calling 'tezos-client' concurrently, force the sequential execution of
'tezos-client' calls by storing them in 'TQueue' and executing items
from this queue in a single thread.