[Chore] Weaken proposal hash lookup condition
Problem: There is an assumption that 'protocolMonitorWorker'
has a bit non-deterministic behavior which can lead to a problem
in case of protocol switch. We can weaken current proposal hash
lookup condition from last block in adoption period to last 2
blocks to make sure that protocol switch will be smooth.

Solution: Weaken the condition for the current proposal hash
lookup and add a TODO to recheck this condition later and make
it more strict.