[#148] Clear ledger queue before starting baking
Problem: From time to time, baker daemon clashes with Ledger
connectivity check done by tezos-client during baker setup.
As a result, a ledger disconnection alert is thrown. However, ledger is
actually connected.

Solution: At the moment, when baker process is about to start, the only
possible queries in the ledger queue are connectivity checks, that are
safe to drop. So the query queue is cleared before starting the baker
daemon. Thus, it doesn't clash with 'tezos-client' call.