[#142] Display latest head level and ledger HWM on the alert
Problem: The description of 'Baker needs to reset HWM' alert
is not informative enough: it doesn't contain information about
latest head level and ledger's high-watermark value. We need to
display there values in the alert description to make it more
informative.

Solution: Display latest head level and ledger high-watermark
on the alert.