[Chore] Catch error in node alert monitor
Problem: After deleting the synced node and creating new one,
the new bootstrapping node starts listening '/monitor/heads'
endpoint instead of using polling. It happens because
'nodeAlertMonitor' function tries to fetch the latest head
using the 'latestFinalHead' TVar and fails with error because
final head's level is higher than the head of the bootstrapping
node. In this case the new node doesn't know about the latest
head because it isn't bootstrapped yet, so we need to catch
that error and set latest head's value to 'Nothing' to make
the alerts mechanism work correctly and node to chose the right
bootstrapping method.

Solution: Catch the RPC error in the 'nodeAlertMonitor' function
and set the latest head value to 'Nothing' in this case.