[#163] Remove unused bash and nix scripts
Problem: The repository contains some outdated
`.sh` and `.nix` files that can be safely deleted.

Solution: Delete the following things:

* `kiln-confugure-tezos.sh`: this script downloads
node and client binaries and starts the Kiln service,
but now we download the binaries using the
`serokell-tezos-binaries/default.nix` and
`brew-tezos-binaries/default.nix`.

* `tests/` folder was deleted because these tests for
accusations and voting are outdated. We have new tests
for accusations in `backend/test/Accusation.hs` and
voting scenario for testing the voting process.

* `install-nix/` folder was deleted because it installs
outdated nix version and there is other ways to install
nix package manager.

* `dep/build-serokell-tezos-binaries/` folder was removed
because it's unused.