[#177] Don't ask user to verify snapshot from xtz-shots
Problem: When uploads/downloads the snapshot, Kiln asks
to verify the head block of this snapshot in the block
explorer to ensure that this is a valid snapshot. But
there is no need to ask user to verify the snapshot
downloaded from xtz-shots metadata because snapshot's
head block hash is known, and 'octez-node' can do the
verification on its own with '--block' option

Solution: Don't ask user to verify the snapshot if this
snapshot was downloaded from xtz-shots metadata. Provide
the head block hash from metadata to 'octez-node' with
'--block' option.