[#177] Fetch snapshot info from xtz-shots metadata
Problem: Kiln parses the 'octez-node snapshot info'
command's stdout to fetch the snapshot's head block
hash, level and timestamp and store them in the db.
But in case when the snapshot was downloaded from
xtz-shots metadata, there is no need to do it because
this info is available in the metadata as well.

Solution: Don't parse the snapshot info from stdout
if this info is already known.