[Chore] Clear 'BakerVote' table in non-voting periods
Problem: Kiln baker's vote in exploration and promotion
periods is stored in the 'BakerVote' table. Currently
this table is cleaned if baker didn't vote in the current
voting period (IOW, if '/votes/ballot_list' endpoint's
response doesn't contain Kiln baker's ballot).

But ballots aren't included to blockhain immediately and
it's needed to wait a couple of blocks to see the ballot
in the '/votes/ballot_list' endpoint. It leads to the
problem that 'BakerVote' table is cleaned before the ballot
is included to the blockchain, which leads to inconsistent
db state.

Solution: Clean 'BakerVote' table only in cooldown and
adoption periods.