[Chore] Don't show 'vote' button if baker has already voted
Problem: Currently Kiln UI allows user to vote more than
one time in exploration and promotion periods. It leads
to the error in the 'octez-client' call and the following
error on the voting prompt:

"The ledger prompt was rejected or timed out. Please try
again".

Solution: Hide the 'Vote' button in exploration and
promotion periods if baker has already voted.