[Chore] Skip voting reminder at the latest block of proposal period
Problem: Kiln can report the voting reminder alerts
(e.g when proposal period ends soon, but user didn't
vote in it). But at the latest block of proposal period
it always return the "You have not voted in the proposal"
period regardless of whether the user voted. It's caused
by the fact that the node rpc returns empty proposal list
if it's queried at the latest block of proposal period.

Solution: Don't report the voting reminder alerts at the
latest block of proposal period.