[Chore] Don't parse metadata if status isn't 200
Problem: We expect 200 response status code when
downloading the metadata from xtz-shots.io. In
case when the status isn't 200, we can't assume
that the response body is a valid metadata json
and there is no much sense to parse it.

Solution: Don't parse metadata if status code isn't
equal to 200.