[Chore] Fix ledger disconnection indicator
Problem: Kiln has a ledger disconnection indicator
on the top bar, which shows that the ledger is
connected when 'ConnectedLedger' table is not empty.
But since ledger connectivity checks were removed,
the table isn't empty only while interactions with
ledger (e.g voting).

Solution: Make this indicator depend on alerts and
show that ledger is connected if there are no
unresolved ledger disconnection alerts in the db.