[Chore] Fix 'protocolMonitorWorker'
Problem: 'protocolMonitorWorker', which is responsible
for starting the baker for the upcoming protocol, wakes up
several times per cycle: at the blocks N-1, N, and N+1
assuming that N is the last block in cycle.

This worker starts the new baker if the current voting
period is 'Adoption' and it's less than 100 blocks
remaining in it. However, given the fact above, this
condition is unreachable on the 'mainnet' and testnets.

Solution: Start the baker for the upcoming protocol
as soon as possible: when the proposal hash is known
and Kiln has the needed binary. It makes this logic
more stable and doesn't affect resources consumption
much since this binary will be idle all the time until
the protocol activation.