[#188] Fix balance display on 'Select address' tab
Problem: We query the ledger to import public key hashes
asynchronously using the 'ledgerIOQueue', but fetch the
balances for them synchronously just after sending the
request to ledger queue. It leads to the problem when
balances are not displayed on UI because the request
in ledger queue isn't finished at the moment of balance
fetching.

Solution: Fetch balance for public key hash just after
importing within the ledger queue request.