[#186] Make ledger polling configurable
Problem: Since concurrent ledger interactions can be
potentially dangerous and may lead to missed opportunities,
it would be nice to allow users to disable the ledger polling
in case when something went wrong.

Solution: Make ledger polling configurable with
'check-ledger-connection' option and enable it by default.
Also don't show the ledger indicator on the header if this
option is disabled and baker doesn't have rights.