[#193] Fix Baked timestamp
Problem: 'Baked: ...' text on the node tile
is redrawn in the following cases:

* When the value inside 'timer' changes (each ~1 sec)
* When node receives new head block, and its timestamp
  changes.

The situation when the latter case appears between the
timer ticks lead to the flickering time format changes
because at that moment timer shows the time which is
less that head's timestamp, so Kiln treats it as a
future timestamp and renders it as 'Baked: Today @ ...'
format.

Solution: Redraw this text only when timer ticks and use
the current value of head's timestamp and this moment.