[#198] Fix styles in 'Add node' tab
Problem: Text inputs in 'Add node' window have
a smaller height than other ones. It's caused by
extra styles for inputs inside the 'fakeRadioButton'
component that is used in 'Settings' tab.

Solution: Add a separate css class for the inputs
in settings tab so they don't overlap with the
default styles in Kiln.