[#197] Optimize download progress UI updates
Problem: Currently Kiln sends updated 'SnapshotMeta'
data each time it receives new chunk of data while
dowloading the snapshot. On the large snapshots
(e.g. on 'ghostnet' or 'mainnet') most of the
updates are redundant because the percentage doesn't
update.

Solution: Send an update to UI only when the downloading
progress percentage, which is displayed on UI, is updated.