[#205] Improve error logs for snapshot import by file path
Problem: There are 2 problems with "Provide path to snapshot
file" option in "Start a Kiln Node" dialog:

1. When snapshot file validation fails, Kiln uses 'logInfo',
but not 'logError', which is misleading.

2. Kiln shows "Invalid snapshot file" error text for any
error in 'octez-node snapshot info' command, which may be
misleading in case of unexpected failures.

Solution:

1. Use 'logError' to log the command's stderr.

2. Show "Invalid snapshot file" message only if
'snapshot info' command failed because the provided
file is invalid. In other cases show the generic error
message.