[#207] Enhance the algorithm of picking the compatible snapshot
Problem: Currently Kiln can choose the snapshot from
snapshot provider metadata that is incompatible with
the Kiln Node version.

Solution: Enhance the algorithm of picking the last
compatible snapshot from the provider metadata and
make Kiln do the following:

1. Try to find the snapshot made by exactly the same
version of 'octez-node' that is used by Kiln.

2. If there is none, try to find the snapshot made
by the same major version of 'octez-node' and less
minor version with 'snapshot_version' equal to the
snapshot version which is supported by current version
of Kiln node (currently set to 5).

3. If there is none, try to find the snapshot with
the 'snapshot_version' which is supported by Kiln
node.