[Chore] Make snapshot metadata tests independent from the current compatible snapshot version
Problem: Currently snapshot metadata tests depend on
the 'compatibleSnapshotVersion' value that is defined
as a constant in the module scope, which leads these
tests to fail when this value is updated.

Solution: Pass needed snapshot version as an argument
to the 'findLatestCompatibleSnapshot' function to make
these tests independent from the value which is currently
set.