[#216] Fix voting on a fresh Kiln install
Problem: When user launched a fresh Kiln install
towards to the end of amendment process, the voting
will not work because of the following reasons:

1. Database tables for the last 4 voting periods
have the foreign key on 'PeriodProposal' table.

2. Since Kiln node is running in a rolling mode,
Kiln can't gather the info about proposal period,
thus 'PeriodProposal' and all other tables will
be empty, which will lead to the broken app state.

Solution: Remove foreign key on 'PeriodProposal'
table in other voting period db tables and store
the proposal hash (which is available from the head
block) and chain_id instead.