[Chore] Weaken the restrictions in the snapshot selection algorithm
Problem: The restrictions in the current implementation
of snapshot selection are sometimes too strict: at the
last step of the algorithm, we try to find the snapshot
with exact same 'snapshot_version'.

Such a restriction leads to the situation when the
algorithm can't find the compatible snapshot for 'nairobinet'
in xtz-shots provider metadata because there are only snapshot
of version 5 for this network.

We can weaken this condition to make our algorithm
handle such cases.

Solution: Changle condition from

```
snapshot.snapshot_version == compatibleSnapshotVersion
```

to

```
compatibleSnapshotVersion - snapshot.snapshot_version <= 1
```