[#61] Optimize alerts storage format
Problem: When missed bake/endorsement is occured,
we insert new row in the db, and pass the list of
alerts to the frontend. In case of many unresolved
alerts, it may be inefficient and leads to UI hanging.

Solution: Rework 'ErrorLogMissedBake' alerts storage
format to store multiple alerts in a single db row.