[Chore] Display original URL when snapshot download failed
Problem: Before displaying the snapshot download failure
error message on UI, Kiln tries to download the snapshot
a couple of times more (because the URL field on UI is
overloaded and it can either be the direct snapshot URL
or the URL to metadata file). When all of these attempts
fail, Kiln shows the modified URL instead of original
in the error message on UI.

Solution: Display the original URL on UI by using
'OverrideUrl' policy.