[#221] Handle missing 'preserved_cycles' in protocol constants
Problem: As per the new protocol changes

https://tezos.gitlab.io/protocols/019_paris.html#protocol-parameters

The `preserved_cycles` field has been replaced with a bunch of other
fields. Among these the `consensus_righs_delay` fields appear to be
usable instead of the `preserved_cycles` field, see,

https://gitlab.com/tezos/tezos/-/issues/6693

So use this field as an alternative, if the `preserved_cycles` field is
missing.