[#215] Add alert when baker doesn't have enough staked balance
Problem: When adaptive issuance will be activated in ParisB
protocol, it's required to have at least 600tz of staking
balance apart from 6000tz total balance to obtain baking
rights. For the bakers which are registered earlier, the
funds will be staked automatically, but this is not the
case for the new bakers. We need to notify users if their
baker doesn't have enough staked balance to obtain voting
rights and advice them to stake funds using Kiln UI.

Solution: Implement the alert that appears if baker's
staked balance if less than required threshold to
get the baking rights. This alert suggests users to
stake funds to get rights.