[#215] Show 'stake' button only when adaptive issuance is activated
Problem: The 'stake' and other operations will be available not
after ParisB activation, but after activation of adaptive issuance
feature (which will be approx. 5 cycles after protocol activation).
We need to show 'stake' button only when adaptive issuance is activated.

Solution: Store adaptive issuance activation cycle fetched from RPC
in 'NodeDataSource', add appropriate function to access it on frontend,
and implement the conditional rendering of 'stake' button in the baker
menu based on it.